(BSR)[PRO] refactor: move Callout to ui kit and move implementations where they are used
