+++ pro/src/components/IndividualOffer/IndivualOfferLayout/OfferStatusBanner/OfferStatusBanner.tsx
@@ -2,8 +2,8 @@ import React from 'react'
 
 import { OfferStatus } from 'apiClient/v1'
 import { CGU_URL } from 'commons/utils/config'
-import { Callout } from 'components/Callout/Callout'
-import { CalloutVariant } from 'components/Callout/types'
+import { Callout } from 'ui-kit/Callout/Callout'
+import { CalloutVariant } from 'ui-kit/Callout/types'
 
 import styles from './OfferStatusBanner.module.scss'
 interface OfferStatusBannerProps {
