+++ pro/src/components/IndividualOffer/SynchronisedProviderInfos/SynchronizedProviderInformation.tsx
@@ -1,6 +1,6 @@
 import { getProviderInfo } from 'commons/core/Providers/utils/getProviderInfo'
-import { Callout } from 'components/Callout/Callout'
-import { CalloutVariant } from 'components/Callout/types'
+import { Callout } from 'ui-kit/Callout/Callout'
+import { CalloutVariant } from 'ui-kit/Callout/types'
 
 import styles from './SynchronizedProviderInformation.module.scss'
