+++ pro/src/components/SignupJourneyForm/Offerer/BannerInvisibleSiren/BannerInvisibleSiren.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
-import { Callout } from 'components/Callout/Callout'
-import { CalloutVariant } from 'components/Callout/types'
+import { Callout } from 'ui-kit/Callout/Callout'
+import { CalloutVariant } from 'ui-kit/Callout/types'
 
 import styles from './BannerInvisibleSiren.module.scss'
