+++ pro/src/pages/Desk/Desk.tsx
@@ -5,8 +5,8 @@ import { apiContremarque } from 'apiClient/api'
 import { isErrorAPIError } from 'apiClient/helpers'
 import { GetBookingResponse } from 'apiClient/v2'
 import { Layout } from 'app/App/layout/Layout'
-import { Callout } from 'components/Callout/Callout'
 import { Button } from 'ui-kit/Button/Button'
+import { Callout } from 'ui-kit/Callout/Callout'
 import { TextInput } from 'ui-kit/form/TextInput/TextInput'
 
 import { BookingDetails } from './BookingDetails'
