+++ pro/src/pages/Homepage/components/AddBankAccountCallout/AddBankAccountCallout.spec.tsx
@@ -6,7 +6,7 @@ import * as useAnalytics from 'app/App/analytics/firebase'
 import { BankAccountEvents } from 'commons/core/FirebaseEvents/constants'
 import { defaultGetOffererResponseModel } from 'commons/utils/factories/individualApiFactories'
 import { renderWithProviders } from 'commons/utils/renderWithProviders'
-import { AddBankAccountCallout } from 'components/Callout/AddBankAccountCallout'
+import { AddBankAccountCallout } from 'pages/Homepage/components/AddBankAccountCallout/AddBankAccountCallout'
 
 const mockLogEvent = vi.fn()
