+++ pro/src/pages/Homepage/components/BankAccountHasPendingCorrectionCallout/BankAccountHasPendingCorrectionCallout.tsx
@@ -3,8 +3,8 @@ import { useLocation } from 'react-router-dom'
 import { GetOffererResponseModel } from 'apiClient/v1'
 import { useAnalytics } from 'app/App/analytics/firebase'
 import { BankAccountEvents } from 'commons/core/FirebaseEvents/constants'
-import { Callout } from 'components/Callout/Callout'
-import { CalloutVariant } from 'components/Callout/types'
+import { Callout } from 'ui-kit/Callout/Callout'
+import { CalloutVariant } from 'ui-kit/Callout/types'
 
 export interface BankAccountHasPendingCorrectionCalloutProps {
   offerer?: GetOffererResponseModel | null
