+++ pro/src/pages/Homepage/components/LinkVenueCallout/LinkVenueCallout.spec.tsx
@@ -1,15 +1,12 @@
 import { screen } from '@testing-library/react'
 import { userEvent } from '@testing-library/user-event'
-import React from 'react'
 
 import * as useAnalytics from 'app/App/analytics/firebase'
 import { BankAccountEvents } from 'commons/core/FirebaseEvents/constants'
 import { defaultGetOffererResponseModel } from 'commons/utils/factories/individualApiFactories'
 import { renderWithProviders } from 'commons/utils/renderWithProviders'
-import {
-  LinkVenueCallout,
-  LinkVenueCalloutProps,
-} from 'components/Callout/LinkVenueCallout'
+
+import { LinkVenueCalloutProps, LinkVenueCallout } from './LinkVenueCallout'
 
 const mockLogEvent = vi.fn()
