+++ pro/src/pages/Homepage/components/Offerers/components/OffererBanners/OffererBanners.tsx
@@ -2,8 +2,8 @@ import React from 'react'
 
 import { GetOffererResponseModel } from 'apiClient/v1'
 import { useActiveFeature } from 'commons/hooks/useActiveFeature'
-import { Callout } from 'components/Callout/Callout'
-import { CalloutVariant } from 'components/Callout/types'
+import { Callout } from 'ui-kit/Callout/Callout'
+import { CalloutVariant } from 'ui-kit/Callout/types'
 
 import { hasOffererAtLeastOnePhysicalVenue } from '../VenueList/venueUtils'
