+++ pro/src/pages/IndividualOffer/IndividualOfferDetails/components/EanSearchCallout/EanSearchCallout.tsx
@@ -1,5 +1,5 @@
-import { Callout } from 'components/Callout/Callout'
-import { CalloutVariant } from 'components/Callout/types'
+import { Callout } from 'ui-kit/Callout/Callout'
+import { CalloutVariant } from 'ui-kit/Callout/types'
 
 import styles from './EanSearchCallout.module.scss'
