+++ pro/src/pages/Offerers/Offerer/VenueV1/fields/PricingPoint/PricingPoint.tsx
@@ -8,12 +8,12 @@ import { GetOffererResponseModel, GetVenueResponseModel } from 'apiClient/v1'
 import { SENT_DATA_ERROR_MESSAGE } from 'commons/core/shared/constants'
 import { useActiveFeature } from 'commons/hooks/useActiveFeature'
 import { useNotification } from 'commons/hooks/useNotification'
-import { Callout } from 'components/Callout/Callout'
 import { ConfirmDialog } from 'components/Dialog/ConfirmDialog/ConfirmDialog'
 import fullLinkIcon from 'icons/full-link.svg'
 import strokeValidIcon from 'icons/stroke-valid.svg'
 import { Button } from 'ui-kit/Button/Button'
 import { ButtonLink } from 'ui-kit/Button/ButtonLink'
+import { Callout } from 'ui-kit/Callout/Callout'
 import { Select } from 'ui-kit/form/Select/Select'
 import { SvgIcon } from 'ui-kit/SvgIcon/SvgIcon'
