+++ pro/src/pages/Reimbursements/BankInformations/PricingPointDialog/PricingPointDialog.tsx
@@ -5,9 +5,9 @@ import { api } from 'apiClient/api'
 import { ManagedVenues } from 'apiClient/v1'
 import { useActiveFeature } from 'commons/hooks/useActiveFeature'
 import { useNotification } from 'commons/hooks/useNotification'
-import { Callout } from 'components/Callout/Callout'
 import { Button } from 'ui-kit/Button/Button'
 import { ButtonVariant } from 'ui-kit/Button/types'
+import { Callout } from 'ui-kit/Callout/Callout'
 import { Select } from 'ui-kit/form/Select/Select'
 
 import styles from './PricingPointDialog.module.scss'
