+++ pro/src/pages/Reimbursements/ReimbursementsInvoices/BannerReimbursementsInfo.tsx
@@ -1,6 +1,4 @@
-import React from 'react'
-
-import { Callout } from 'components/Callout/Callout'
+import { Callout } from 'ui-kit/Callout/Callout'
 
 export const BannerReimbursementsInfo = (): JSX.Element => {
   return (
