+++ pro/src/pages/VenueCreation/CollectiveDmsTimeline/CollectiveDmsTimeline.tsx
@@ -2,10 +2,10 @@ import { DMSApplicationForEAC, DMSApplicationstatus } from 'apiClient/v1'
 import { useAnalytics } from 'app/App/analytics/firebase'
 import { Events } from 'commons/core/FirebaseEvents/constants'
 import { getDateToFrenchText } from 'commons/utils/date'
-import { Callout } from 'components/Callout/Callout'
-import { CalloutVariant } from 'components/Callout/types'
 import { ButtonLink } from 'ui-kit/Button/ButtonLink'
 import { ButtonVariant } from 'ui-kit/Button/types'
+import { Callout } from 'ui-kit/Callout/Callout'
+import { CalloutVariant } from 'ui-kit/Callout/types'
 import { Timeline, TimelineStepType } from 'ui-kit/Timeline/Timeline'
 
 import styles from './CollectiveDmsTimeline.module.scss'
