+++ pro/src/pages/VenueEdition/VenueEditionFormScreen.tsx
@@ -2,8 +2,8 @@ import { Route, Routes } from 'react-router-dom'
 
 import { GetVenueResponseModel } from 'apiClient/v1'
 import { useActiveFeature } from 'commons/hooks/useActiveFeature'
-import { Callout } from 'components/Callout/Callout'
 import { PartnerPageIndividualSection } from 'pages/Homepage/components/Offerers/components/PartnerPages/components/PartnerPageIndividualSection'
+import { Callout } from 'ui-kit/Callout/Callout'
 
 import { AccesLibreSection } from './AccesLibreSection/AccesLibreSection'
 import { VenueEditionForm } from './VenueEditionForm'
