+++ pro/src/ui-kit/Callout/Callout.tsx
@@ -1,12 +1,12 @@
 import cn from 'classnames'
 import React from 'react'
 
-import { CalloutVariant } from 'components/Callout/types'
 import fullErrorIcon from 'icons/full-error.svg'
 import fullInfoIcon from 'icons/full-info.svg'
 import fullValidIcon from 'icons/full-validate.svg'
 import fullWarningIcon from 'icons/full-warning.svg'
 import strokeCloseIcon from 'icons/stroke-close.svg'
+import { CalloutVariant } from 'ui-kit/Callout/types'
 import { Link, LinkNodes } from 'ui-kit/LinkNodes/LinkNodes'
 import { SvgIcon } from 'ui-kit/SvgIcon/SvgIcon'
